test the source and not the installed version
As we need to make sure to include to include the relevant 'lib'
directory we change to the 't' directory for the tests.  This is
a common way to solve the problem.

t/pod.t needs to run from the parent directory of 't' so that it
finds all .pod files.